requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
pydantic==1.8.2
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sniffio==1.2.0 
starlette==0.16.0
twilio==7.3.2
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.15.0
python-dotenv == 0.19.2
openai == 0.8.0
python-dotenv == 0.19.2